+++ RELEASE-NOTES.md
@@ -4,6 +4,7 @@
 * #97 Support JAX-RS multipart out of the box
 * #98 Upgrade to JAX-RS 4.0.0 and Jersey 4.0.0
 * #99 Upgrade Wiremock to 4.0.0-beta.24
+* #100 Disable WADL feature by default
 
 ## 4.0-M1